refactor: extract info component overrides to oas31 plugin
