+++ src/core/plugins/oas3/wrap-components/index.js
@@ -4,7 +4,6 @@ import VersionStamp from "./version-stamp"
 import OnlineValidatorBadge from "./online-validator-badge"
 import Model from "./model"
 import JsonSchema_string from "./json-schema-string"
-import info from "./info"
 
 export default {
   Markdown,
@@ -13,5 +12,4 @@ export default {
   VersionStamp,
   model: Model,
   onlineValidatorBadge: OnlineValidatorBadge,
-  info,
 }
