+++ src/core/plugins/oas31/wrap-components/license.jsx
@@ -2,6 +2,7 @@
  * @prettier
  */
 import React from "react"
+
 const LicenseWrapper = (Original, system) => (props) => {
   if (system.specSelectors.isOAS31()) {
     const OAS31License = system.getComponent("OAS31License")
